pyproject.toml
[project]
name = "russian-roulette-bot"
version = "0.1.0"
description = "A Discord bot to play virtual Russian Roulette with your friends."
readme = "readme.md"
requires-python = ">=3.11"
dependencies = [
    "deta~=1.1.0",
    "discord-py~=2.4.0",
    "python-dotenv~=1.0.0",
    "pyyaml~=6.0.0",
]

[dependency-groups]
dev = [
    "pre-commit~=3.8.0",
]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.ruff]
src = ["src"]
line-length = 119
target-version = "py311"

[tool.ruff.lint]
select = ["ALL"]
ignore = ["A", "D", "T20", "TD002", "TD003"]

[tool.pyright]
venvPath = "."
venv = ".venv"
